+++ src/src/registration_impl.cpp
@@ -11,14 +11,14 @@ namespace lidar_utils {
 namespace internal {
 
 void executeScanToMapMatching(CloudType::Ptr &cloud, CloudType::Ptr &map,
-                              Eigen::Matrix4f &in_transform,
-                              Eigen::Matrix4f &out_transform,
+                              Eigen::Matrix4d &in_transform,
+                              Eigen::Matrix4d &out_transform,
                               float max_correspondence_distance,
                               float voxel_size, float score_threshold) {}
 
 void executeScanToScanMatching(CloudType::Ptr &cloud, CloudType::Ptr &localmap,
-                               Eigen::Matrix4f &in_transform,
-                               Eigen::Matrix4f &out_transform,
+                               Eigen::Matrix4d &in_transform,
+                               Eigen::Matrix4d &out_transform,
                                float max_correspondence_distance,
                                float voxel_size, float score_threshold) {}
